Implemented Service for Searching for Events within Date Range, Location Bounding Box and by Keyword

+++ TE.Logic.Events/Services/TrainingEvents/Dtos/GeoLocationDto.cs
@@ -2,7 +2,7 @@ namespace TE.Logic.Events.Services.TrainingEvents.Dtos
 {
     public class GeoLocationDto
     {
-        public long Latitude { get; set; }
-        public long Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
     }
 }
